refactor cart to separate itemlist, quantity, remove btn
