+++ client/src/App.tsx
@@ -1,4 +1,6 @@
+import CheckoutPage from "./pages/checkout/CheckoutPage";
+
 export default function App() {
-  return <div></div>;
+  return <div><CheckoutPage/></div>;
 }
