+++ client/src/pages/checkout/CheckoutPage.tsx
@@ -1,0 +1,8 @@
+import ItemList from "../../components/ItemList";
+
+
+export default function Checkout() {
+  return <ItemList />;
+{/* <Prices/>
+<SubmitButton/> */}
+}
